Bookshop/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.18
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-countries==7.1
django-crispy==1.8.1
django-mathfilters==1.0.0
et-xmlfile==1.0.1
idna==2.10
kombu==3.0.37
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==2.10.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.57.0
urllib3==1.26.4
xlrd==2.0.1
